refactor: useCVC hook으로 분리
